requirements.txt
playwright
allure-pytest
pytest
